requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
pypinyin
jieba
deep-translator
